XiYan-SQLTraining/requirements.txt
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1
numpy==1.26.4
accelerate==0.31.0
deepspeed==0.12.0
transformers==4.42.3
peft==0.11.1
flash_attn==2.5.9.post1
datasets==2.18.0
protobuf==5.27.2
sentencepiece==0.2.0
swanlab==0.6.0
modelscope==1.26.0
psycopg2-binary==2.9.9
mysql-connector==2.2.9
mysql-connector-python==9.0.0
sqlglot==26.16.2
textdistance==4.6.3
pandas==2.2.3
SQLAlchemy==2.0.16
llama-index==0.11.12
llama-index-llms-dashscope